# swagger_full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flash API",
    "version": "1.0.0",
    "description": "API completa para gerenciamento de múltiplas sessões WhatsApp",
    "contact": {
      "name": "API Support",
      "email": "quickgestor@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de Beta"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "API Key para autenticação"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/chat/send-text": {
      "post": {
        "summary": "Enviar mensagem de texto",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "text"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Numero de quem vai receber a mensagem",
                    "default": "5599999999999"
                  },
                  "text": {
                    "type": "string",
                    "description": "Texto a ser enviado",
                    "default": "Olá mundo"
                  },
                  "linkPreview": {
                    "type": "boolean",
                    "description": "false ou true",
                    "default": "false"
                  },
                  "mentions": {
                    "type": "array",
                    "description": "[ '5599999999999']",
                    "items": {
                      "type": "string"
                    }
                  },
                  "delay": {
                    "type": "number",
                    "description": "Delay de mensagem",
                    "default": "1200"
                  },
                  "useQueue": {
                    "type": "boolean",
                    "description": "Adicionar mensagem em fila",
                    "default": "false"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mensagem enviada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Mensagem de texto enviada com sucesso"
                    },
                    "data": {
                      "type": "array",
                      "example": "['key[]', 'message[]', 'messageTimestamp', 'status']"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/send-image": {
      "post": {
        "summary": "Enviar imagem",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "caption",
                  "image"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Numero de quem vai receber a mensagem",
                    "default": "5599999999999"
                  },
                  "caption": {
                    "type": "string",
                    "description": "Texto a ser enviado",
                    "default": "Olá mundo"
                  },
                  "image": {
                    "type": "string",
                    "description": "Link ou base64",
                    "default": "https://exemplo.com.br/imagem.png"
                  },
                  "mentions": {
                    "type": "array",
                    "description": "[ '5599999999999']",
                    "items": {
                      "type": "string"
                    }
                  },
                  "delay": {
                    "type": "number",
                    "description": "Delay de mensagem",
                    "default": "1200"
                  },
                  "useQueue": {
                    "type": "boolean",
                    "description": "Adicionar mensagem em fila",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "imagem enviada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "imagem com imagem enviada com sucesso"
                    },
                    "data": {
                      "type": "array",
                      "example": "['key[]', 'message[]', 'messageTimestamp', 'status']"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/send-video": {
      "post": {
        "summary": "Enviar Video",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "caption",
                  "video"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Numero de quem vai receber a mensagem",
                    "default": "5599999999999"
                  },
                  "caption": {
                    "type": "string",
                    "description": "Texto a ser enviado",
                    "default": "Olá mundo"
                  },
                  "video": {
                    "type": "string",
                    "description": "Link ou base64",
                    "default": "https://exemplo.com.br/video.mp4"
                  },
                  "gifPlayback": {
                    "type": "boolean",
                    "description": "Enviar em modo gif",
                    "default": false
                  },
                  "delay": {
                    "type": "number",
                    "description": "Delay de mensagem",
                    "default": "1200"
                  },
                  "useQueue": {
                    "type": "boolean",
                    "description": "Adicionar mensagem em fila",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mensagem enviada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Mensagem com imagem enviada com sucesso"
                    },
                    "data": {
                      "type": "array",
                      "example": "['key[]', 'message[]', 'messageTimestamp', 'status']"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/chat/send-audio": {
      "post": {
        "summary": "Enviar áudio",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "to",
                  "audio"
                ],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Número do destinatário",
                    "default": "5599999999999"
                  },
                  "audio": {
                    "type": "string",
                    "description": "URL do áudio ou base64",
                    "default": "https://exemplo.com.br/video.mp3"
                  },
                  "ptt": {
                    "type": "boolean",
                    "description": "Enviar como nota de voz",
                    "default": false
                  },
                  "delay": {
                    "type": "number",
                    "description": "Delay em milissegundos",
                    "default": 1200
                  },
                  "useQueue": {
                    "type": "boolean",
                    "description": "Usar fila de mensagens",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Áudio enviado com sucesso"
          }
        }
      }
    },
...
    "/chat/messages": {
      "get": {
        "summary": "Obter mensagens do chat",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "jid",
            "schema": {
              "type": "string"
            },
            "description": "JID do chat (opcional)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Limite de mensagens"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Offset para paginação"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mensagens"
          }
        }
      }
    },
    "/chat/chats": {
      "get": {
        "summary": "Obter lista de chats",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de chats"
          }
        }
      }
    },
    "/chat/queue/status": {
      "get": {
        "summary": "Obter status da fila de mensagens",
        "tags": [
          "Chats"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Status da fila"
          }
        }
      }
    },
...
  }
}